Reorganized testclasses
